WIP: testing out moving the load_and_broadcast to happen in the websocket part only if resource has subscribers.
