+++ services/ui_backend_service/api/notify.py
@@ -1,17 +1,15 @@
 import json
 import asyncio
-from typing import Dict, List, Any, Callable
+from typing import Dict, List
 from services.data.postgres_async_db import AsyncPostgresDB
-from pyee import AsyncIOEventEmitter
-from .data_refiner import TaskRefiner
+from pyee import ExecutorEventEmitter
 
 
 class ListenNotify(object):
     def __init__(self, app, event_emitter=None):
         self.app = app
-        self.event_emitter = event_emitter or AsyncIOEventEmitter()
+        self.event_emitter = event_emitter or ExecutorEventEmitter()
         self.db = AsyncPostgresDB.get_instance()
-        self.task_refiner = TaskRefiner()
 
         self.loop = asyncio.get_event_loop()
         self.loop.create_task(self._init())
@@ -41,12 +39,11 @@ class ListenNotify(object):
             table = await self.db.get_table_by_name(table_name)
             if table != None:
                 resources = resource_list(table.table_name, data)
-                postprocess = self.get_table_postprocessor(table.table_name)
+
                 # Broadcast this event to `api/ws.py` (Websocket.event_handler)
                 # and notify each Websocket connection about this event
                 if resources != None and len(resources) > 0:
-                    await load_and_broadcast(self.event_emitter, operation, table,
-                                                data, table.primary_keys, postprocess=postprocess)
+                    await _broadcast(self.event_emitter, operation, table, data)
 
                 # Heartbeat watcher for Runs.
                 if table.table_name == self.db.run_table_postgres.table_name:
@@ -76,22 +73,19 @@ class ListenNotify(object):
                     # First attempt has already been inserted by task table trigger.
                     # Later attempts must count as inserts to register properly for the UI
                     _op = "UPDATE" if _attempt_id == 0 else "INSERT"
-                    await load_and_broadcast(
+                    await _broadcast(
                         event_emitter=self.event_emitter,
                         operation=_op,
                         table=self.db.task_table_postgres,
-                        data=data,
-                        keys=self.db.task_table_postgres.primary_keys,
-                        custom_filter_dict={"attempt_id": _attempt_id},
-                        postprocess=self.task_refiner.postprocess
+                        data={**data, "attempt_id": _attempt_id}
                     )
 
                     # Last step is always called 'end' and only one '_task_ok' should be present
                     # Run is considered finished once 'end' step has '_task_ok' artifact
                     if data["step_name"] == "end":
-                        await load_and_broadcast(
+                        await _broadcast(
                             self.event_emitter, "UPDATE", self.db.run_table_postgres,
-                            data, self.db.run_table_postgres.primary_keys)
+                            data)
                         # Also trigger preload of artifacts after a run finishes.
                         self.event_emitter.emit("preload-artifacts", data['run_number'])
                         # And remove possible heartbeat watchers for completed runs
@@ -99,13 +93,6 @@ class ListenNotify(object):
 
         except Exception as err:
             print(err, flush=True)
-
-    def get_table_postprocessor(self, table_name):
-        if table_name == self.db.task_table_postgres.table_name:
-            return self.task_refiner.postprocess
-        else:
-            return None
-
 
 def resource_list(table_name: str, data: Dict):
     resource_paths = {
@@ -127,43 +114,12 @@ def resource_list(table_name: str, data: Dict):
             "/flows/{flow_id}/runs/{run_number}/steps/{step_name}/tasks",
             "/flows/{flow_id}/runs/{run_number}/steps/{step_name}/tasks/{task_id}",
             "/flows/{flow_id}/runs/{run_number}/steps/{step_name}/tasks/{task_id}/attempts"
-        ],
-        # "metadata_v3": [
-        #     "/flows/{flow_id}/runs/{run_number}/metadata",
-        #     "/flows/{flow_id}/runs/{run_number}/steps/{step_name}/tasks/{task_id}/metadata"
-        # ],
-        # "artifact_v3": [
-        #     "/flows/{flow_id}/runs/{run_number}/artifacts",
-        #     "/flows/{flow_id}/runs/{run_number}/steps/{step_name}/artifacts",
-        #     "/flows/{flow_id}/runs/{run_number}/steps/{step_name}/tasks/{task_id}/artifacts"
-        # ],
+        ]
     }
     if table_name in resource_paths:
         return [path.format(**data) for path in resource_paths[table_name]]
     return []
 
-
-async def load_data_from_db(table, filter_dict: Dict[str, Any],
-                            postprocess: Callable = None):
-    conditions, values = [], []
-    for k, v in filter_dict.items():
-        conditions.append("{} = %s".format(k))
-        values.append(v)
-
-    results, _ = await table.find_records(
-        conditions=conditions, values=values, fetch_single=True,
-        enable_joins=True, postprocess=postprocess
-    )
-    resources = resource_list(table.table_name, results.body)
-    return results.body, resources
-
-
-async def load_and_broadcast(event_emitter, operation: str, table,
-                             data: Dict, keys: List[str], custom_filter_dict={},
-                             postprocess: Callable = None):
-    filter_dict = {}
-    for k in keys:
-        filter_dict[k] = data[k]
-
-    _data, _resources = await load_data_from_db(table, {**filter_dict, **custom_filter_dict}, postprocess)
-    event_emitter.emit('notify', operation, _resources, _data)
+async def _broadcast(event_emitter, operation: str, table, data: Dict):
+    _resources = resource_list(table.table_name, data)
+    event_emitter.emit('notify', operation, table, _resources, data)
